requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
aiohttp==3.9.1
loguru==0.7.2
schedule==1.2.1
rich==13.7.0
plotly==5.18.0
click==8.1.7
websocket-client==1.6.4
